functions/requirements.txt
google-cloud-storage
google-cloud-firestore
openpyxl